Убрал черезмерное уменьшение компонентов
