--- project/src/components/btn-add-list/btn-add-list.tsx
@@ -1,12 +1,0 @@
-function BtnAddList ():JSX.Element {
-  return (
-    <button className="btn btn--list film-card__button" type="button">
-      <svg viewBox="0 0 19 20" width="19" height="20">
-        <use xlinkHref="#add"/>
-      </svg>
-      <span>My list</span>
-    </button>
-  );
-}
-
-export default BtnAddList;
