--- project/src/components/btn-add-review/btn-add-review.tsx
@@ -1,5 +1,0 @@
-function BtnAddReview ():JSX.Element {
-  return (<a href='add-review.html' className="btn film-card__button">Add review</a>);
-}
-
-export default BtnAddReview;
