--- project/src/components/btn-in-list/btn-in-list.tsx
@@ -1,12 +1,0 @@
-function BtnInList (): JSX.Element {
-  return (
-    <button className="btn btn--list film-card__button" type="button">
-      <svg viewBox="0 0 18 14" width="18" height="14">
-        <use xlinkHref="#in-list"></use>
-      </svg>
-      <span>My list</span>
-    </button>
-  );
-}
-
-export default BtnInList;
