--- project/src/components/catalog-more/catalog-more.tsx
@@ -1,11 +1,0 @@
-import BtnShowMore from '../btn-show-more/btn-show-more';
-
-function CatalogMore ():JSX.Element {
-  return (
-    <div className="catalog__more">
-      <BtnShowMore />
-    </div>
-  );
-}
-
-export default CatalogMore;
