+++ project/src/components/catalog/catalog.tsx
@@ -1,25 +1,20 @@
 import GenresList from '../genres-list/genres-list';
 import CatalogFilmsList from '../catalog-films-list/catalog-films-list';
-import CatalogMore from '../catalog-more/catalog-more';
+import {Film} from '../../types/film';
+import BtnShowMore from '../btn-show-more/btn-show-more';
 
-type DataProps = {
-  id: number,
-  name: string,
-  image: string,
-  description: string,
-};
 
 type CatalogProps = {
-  cardsData: DataProps[]
+  filmCardsData: Film[]
 }
 
-function Catalog ({cardsData}: CatalogProps):JSX.Element {
+function Catalog({filmCardsData}: CatalogProps): JSX.Element {
   return (
     <section className="catalog">
       <h2 className="catalog__title visually-hidden">Catalog</h2>
       <GenresList />
-      <CatalogFilmsList cardsData = {cardsData}/>
-      <CatalogMore />
+      <CatalogFilmsList filmCardsData={filmCardsData} />
+      <BtnShowMore />
     </section>
   );
 }
