--- project/src/components/copyright/copyright.tsx
@@ -1,9 +1,0 @@
-function Copyright ():JSX.Element {
-  return (
-    <div className="copyright">
-      <p>© 2019 What to watch Ltd.</p>
-    </div>
-  );
-}
-
-export default Copyright;
