--- project/src/components/film-card-bg/film-card-bg.tsx
@@ -1,13 +1,0 @@
-type FilmCardBgProps = {
-  bgImage: string,
-}
-
-function FilmCardBg ({bgImage}:FilmCardBgProps):JSX.Element {
-  return (
-    <div className="film-card__bg">
-      <img src={`img/${bgImage}`} alt="The Grand Budapest Hotel"/>
-    </div>
-  );
-}
-
-export default FilmCardBg;
