--- project/src/components/film-card-details/film-card-details.tsx
@@ -1,55 +1,0 @@
-type FilmCardDetailsData = {
-  director: string,
-  actors: string[],
-  runtime: number,
-  genre: string,
-  release: number,
-}
-
-type FilmCardDetailsProps = {
-  filmData: FilmCardDetailsData
-}
-
-function FilmCardDetails ({filmData}: FilmCardDetailsProps): JSX.Element {
-  const {
-    director,
-    actors,
-    runtime,
-    genre,
-    release,
-  } = filmData;
-
-  return (
-    <div className="film-card__text film-card__row">
-      <div className="film-card__text-col">
-        <p className="film-card__details-item">
-          <strong className="film-card__details-name">Director</strong>
-          <span className="film-card__details-value">{director}</span>
-        </p>
-        <p className="film-card__details-item">
-          <strong className="film-card__details-name">Starring</strong>
-          <span className="film-card__details-value">
-            {actors.map((actor) => `${actor}, `)}
-          </span>
-        </p>
-      </div>
-
-      <div className="film-card__text-col">
-        <p className="film-card__details-item">
-          <strong className="film-card__details-name">Run Time</strong>
-          <span className="film-card__details-value">{runtime}</span>
-        </p>
-        <p className="film-card__details-item">
-          <strong className="film-card__details-name">Genre</strong>
-          <span className="film-card__details-value">{genre}</span>
-        </p>
-        <p className="film-card__details-item">
-          <strong className="film-card__details-name">Released</strong>
-          <span className="film-card__details-value">{release}</span>
-        </p>
-      </div>
-    </div>
-  );
-}
-
-export default FilmCardDetails;
