--- project/src/components/film-card-hero/film-card-hero.tsx
@@ -1,40 +1,0 @@
-import FilmCardBg from '../film-card-bg/film-card-bg';
-import Header from '../header/header';
-
-function FilmCardHero ():JSX.Element {
-  return (
-    <div className="film-card__hero">
-      <FilmCardBg bgImage={''} />
-
-      <h1 className="visually-hidden">WTW</h1>
-      <Header />
-
-      <div className="film-card__wrap">
-        <div className="film-card__desc">
-          <h2 className="film-card__title">The Grand Budapest Hotel</h2>
-          <p className="film-card__meta">
-            <span className="film-card__genre">Drama</span>
-            <span className="film-card__year">2014</span>
-          </p>
-
-          <div className="film-card__buttons">
-            <button className="btn btn--play film-card__button" type="button">
-              <svg viewBox="0 0 19 19" width="19" height="19">
-                <use xlinkHref="#play-s"></use>
-              </svg>
-              <span>Play</span>
-            </button>
-            <button className="btn btn--list film-card__button" type="button">
-              <svg viewBox="0 0 19 20" width="19" height="20">
-                <use xlinkHref="#add"></use>
-              </svg>
-              <span>My list</span>
-            </button>
-            <a href="add-review.html" className="btn film-card__button">Add review</a>
-          </div>
-        </div>
-      </div>
-    </div>);
-}
-
-export default FilmCardHero;
