+++ project/src/components/genres-list/genres-list.tsx
@@ -1,4 +1,4 @@
-function GenresList ():JSX.Element {
+function GenresList(): JSX.Element {
   return (
     <ul className="catalog__genres-list">
       <li className="catalog__genres-item catalog__genres-item--active">
