+++ project/src/components/header/header.tsx
@@ -1,11 +1,11 @@
 import Logo from '../logo/logo';
 import UserBlock from '../user-block/user-block';
 
-function Header (): JSX.Element {
+function Header(): JSX.Element {
   return (
     <header className="page-header film-card__head">
-      <Logo modification={false}/>
-      <UserBlock/>
+      <Logo />
+      <UserBlock />
     </header>
   );
 }
