+++ project/src/components/logo/logo.tsx
@@ -1,11 +1,11 @@
 type LogoProps = {
-  modification: boolean,
+  className?: string
 };
 
-function Logo ({modification}: LogoProps):JSX.Element {
+function Logo({className = ''}: LogoProps): JSX.Element {
   return (
     <div className="logo">
-      <a className={`logo__link ${modification ? 'logo__link--light': ''}`} >
+      <a className={`logo__link ${className}`} >
         <span className="logo__letter logo__letter--1">W</span>
         <span className="logo__letter logo__letter--2">T</span>
         <span className="logo__letter logo__letter--3">W</span>
