+++ project/src/types/film.ts
@@ -1,0 +1,14 @@
+export type Film = {
+  id: number,
+  name: string,
+  image: string,
+  description: string,
+}
+
+export type FilmPromo = {
+  image?: string,
+  bgImage?: string,
+  name: string,
+  genre: string,
+  data: number,
+}
